+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [master](https://github.com/arangodb/kube-arangodb/tree/master) (N/A)
+- (Bugfix) (Gateway) Enable Extended CONNECT on the upstream HTTP/2 cluster so WebSocket upgrades work over HTTP/2 upstreams as well as HTTP/1.1
 - (Feature) (Platform) Discover platform service Helm releases by label and create an ArangoPlatformWorkflow for each behind the platform-workflow-discovery feature
 - (Feature) (Platform) Discover and reflect an existing Helm release in the status of a chart-less ArangoPlatformWorkflow
 - (Feature) (Platform) Collect ArangoPlatformWorkflow resources in the debug package